+++ sysdeps/nptl/pthread.h
@@ -1162,7 +1162,7 @@ typedef uint8_t accel_primitive_t; // TODO: This should be defined in a more app
 
 extern int pthread_attr_setprimitive_np(pthread_attr_t *attr, accel_primitive_t prim) __THROW;
 extern int pthread_attr_setmmiobase_np(pthread_attr_t *attr, void *mem, void *buf2handle_list) __THROW;
-extern int pthread_attr_setsmqueue_np(pthread_attr_t *attr, uint64_t queue_ptr) __THROW;
+extern int pthread_attr_setsmqueue_np(pthread_attr_t *attr, unsigned queue_ptr) __THROW;
 
 __END_DECLS
